+++ components/basecomponents/KidsTranslation.tsx
@@ -23,6 +23,7 @@ type KidsTranslationProps = {
   soundUrl: string;
   language: Language;
   isActive?: boolean;
+  isCard?: boolean;
 };
 
 export const KidsTranslation = ({
@@ -31,6 +32,7 @@ export const KidsTranslation = ({
   language,
   soundUrl,
   isActive = false,
+  isCard = false,
 }: KidsTranslationProps): JSX.Element => {
   const [audioPlaying, setIsPlaying] = useAtom(dictionaryAudioPlayAtom);
   const [activePhrase, setActivePhrase] = useAtom(dictionaryActivePhraseAtom);
@@ -47,6 +49,7 @@ export const KidsTranslation = ({
     setIsPlaying(false);
   }, [audioPlaying, soundUrl, setIsPlaying, setActivePhrase, translation]);
 
+  /* Kids Translation as default (web site) */
   const renderDefault = () => {
     return (
       <div className="flex justify-between items-center py-2 ">
@@ -71,6 +74,32 @@ export const KidsTranslation = ({
     );
   };
 
+  /** Kids Translation as card */
+  const renderAsCard = () => {
+    return (
+      <div className="flex justify-between items-center py-3 px-5 bg-white rounded-xl shadow-xl mb-20">
+        <div className="w-full">
+          <div className="flex items-center mb-2 w-56">
+            <Flag language={language} width={30} height={30} className={'mr-3'} />
+            <p>
+              <Trans className="block my-2">{t(`dictionary_page.${language}`)}</Trans>
+            </p>
+          </div>
+          <p className="self-start w-full">{translation}</p>
+          <p className="text-gray-500">{`[\u00A0${transcription}\u00A0]`}</p>
+        </div>
+        <button onClick={handleClick} aria-label={`${t(PLAY_ACTION_TRANSLATION_ID)} ${translation}`}>
+          <PlayKidsIcon
+            className={`cursor-pointer active:scale-75 transition-all duration-300 w-24 ${styles.playIcon} ${
+              activePhrase === translation && audioPlaying ? styles.pulse : ''
+            }`}
+          />
+        </button>
+      </div>
+    );
+  };
+
+  /** Kids Translation for kids kiosk */
   const renderForKiosk = () => {
     const playButtonClasses = `flex grow  ${language === 'uk' ? `bg-[${KIOSK_BG_COLOR_UK}]` : `bg-[${KIOSK_BG_COLOR_OTHER}]`} ${
       isActive ? 'flex-row w-full justify-center items-center h-[122px]' : 'flex-col justify-between items-center py-5 w-1/2'
@@ -87,7 +116,7 @@ export const KidsTranslation = ({
 
   switch (renderFor) {
     case Platform.KIOSK:
-      return renderForKiosk();
+      return isCard ? renderAsCard() : renderForKiosk();
     case Platform.WEB:
     default:
       return renderDefault();
