+++ components/basecomponents/QuizGame/imageStyle.module.css
