+++ data/kiosk-games.ts
@@ -6,6 +6,6 @@ export type Game = {
 export const GAMES: Game[] = [
   { name: 'pexeso', image: 'pexeso.png' },
   { name: 'slovicka', image: 'slovicka.png' },
-  { name: 'kviz', image: 'kviz.png' },
+  { name: 'quiz', image: 'kviz.png' },
   { name: 'stories', image: 'pohadky.png' },
 ];
