+++ pages/kiosk/game/[name].tsx
@@ -3,6 +3,7 @@ import dynamic from 'next/dynamic';
 import { ParsedUrlQuery } from 'querystring';
 import stories from 'data/stories';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
+import { DictionaryDataObject, fetchDictionary } from '../../../utils/getDataUtils';
 
 /** Components */
 const MemoryGame = dynamic(() => import('components/basecomponents/MemoryGame/MemoryGame'), {
@@ -21,6 +22,7 @@ import { getServerSideTranslations } from 'utils/localization';
 /** Data */
 import { GAMES, Game as GameType } from 'data/kiosk-games';
 import { usePlatformDetection } from 'utils/usePlatformDetection';
+import Quiz from 'components/basecomponents/QuizGame/Quiz';
 
 interface GamePageParams extends ParsedUrlQuery {
   name: string;
@@ -29,9 +31,10 @@ interface GamePageParams extends ParsedUrlQuery {
 interface GameProps {
   gameName: string;
   stories: Story[];
+  dictionary: DictionaryDataObject;
 }
 
-const Game = ({ gameName, stories }: GameProps) => {
+const Game = ({ gameName, stories, dictionary }: GameProps) => {
   usePlatformDetection();
 
   switch (gameName) {
@@ -41,6 +44,8 @@ const Game = ({ gameName, stories }: GameProps) => {
       return <KidsDictionary />;
     case 'stories':
       return <StoriesList stories={stories} />;
+    case 'quiz':
+      return <Quiz dictionary={dictionary} />;
     default:
       return <div>Game not implemented yet.</div>;
   }
@@ -63,6 +68,7 @@ export const getStaticPaths: GetStaticPaths<GamePageParams> = async () => {
 
 export const getStaticProps: GetStaticProps = async ({ params, locale }) => {
   const localeTranslations = await getServerSideTranslations(locale);
+  const dictionary = await fetchDictionary();
 
   const gameName = params?.name ?? '';
 
@@ -71,6 +77,7 @@ export const getStaticProps: GetStaticProps = async ({ params, locale }) => {
       ...localeTranslations,
       gameName,
       stories,
+      dictionary,
       ...(await serverSideTranslations(locale ?? 'cs', ['common'])),
     },
   };
